chat-app-master/chat-app-master/back-end/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
Django==5.0.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-staticfiles==1.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
zope.interface==6.1
